Switch to pyproject.toml-based setup
This change allows the project to build compatible sdists and wheels
following modern packaging standards. Declarative metadata is better for
the ecosystem, as it allows tooling to learn about the package without
having to build it first. (Of course, building is not normally required
for this project anyway, since it distributes a wheel.)

Using the new standards, everything can be done without any code in
`setup.py` nor special instructions in `MANIFEST.in`.